Move some code core.py to new modules
